Refactor convolve_2d_scale unit tests
The filter values used in the scaling algorithm are specified in the
documentation of the inter-prediction process (chapter 7.11.3.4. [1])
which contains the scaling algorithm. An important characteristic of
these filter values is that all of them are even and we can leverage
this information when optimizing the av1_convolve_2d_scale function.
However the unit tests do not follow this pattern, the test currently
generates arbitrary filter values instead.

To enable us to take advantage of specific filter characteristics,
adjust the unit tests to use the predefined filter values of
av1_interp_filter_params_list defined in filter.h file.

Additionally, update convolution parameters when testing with different
is_compound values.

[1]https://aomediacodec.github.io/av1-spec/av1-spec.pdf

This is a port of d00a1befd9275278994f670bf2392aa189a649bb in libaom.